+++ packages/core/src/LiveSeries.ts
@@ -20,10 +20,7 @@ import {
   type LiveRollingOptions,
   type RollingWindow,
 } from './LiveRollingAggregation.js';
-import {
-  LiveSequenceRollingAggregation,
-  type SequenceInterval,
-} from './LiveSequenceRollingAggregation.js';
+import { LiveSequenceRollingAggregation } from './LiveSequenceRollingAggregation.js';
 import { TimeSeries } from './TimeSeries.js';
 import { ValidationError } from './errors.js';
 import { parseJsonRows } from './json.js';
@@ -549,7 +546,9 @@ export class LiveSeries<S extends SeriesSchema> {
     windowOrMapping: RollingWindow | M,
     mappingOrOptions?: M | LiveRollingOptions,
     options?: LiveRollingOptions,
-  ): LiveRollingAggregation<S, RollingSchema<S, M>> | LiveSequenceRollingAggregation<S, RollingSchema<S, M>> {
+  ):
+    | LiveRollingAggregation<S, RollingSchema<S, M>>
+    | LiveSequenceRollingAggregation<S, RollingSchema<S, M>> {
     if (sequenceOrWindow instanceof Sequence) {
       const r = new LiveRollingAggregation(
         this,
@@ -557,7 +556,11 @@ export class LiveSeries<S extends SeriesSchema> {
         mappingOrOptions as AggregateMap<S>,
         options,
       );
-      return new LiveSequenceRollingAggregation(r, sequenceOrWindow) as any;
+      return new LiveSequenceRollingAggregation(
+        r,
+        sequenceOrWindow,
+        true,
+      ) as any;
     }
     return new LiveRollingAggregation(
       this,